refactor: Put INIT_DATA files into initialDataType directory
